# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "description": "Uniform error envelope.",
        "examples": [
          {
            "code": "not_found",
            "message": "The requested resource does not exist.",
            "request_id": "9f1c2b7a4e5d4f0aa1b2c3d4e5f60718"
          }
        ],
        "properties": {
          "code": {
            "description": "Stable machine-readable error identifier. Safe to branch on; the human-readable 'message' is not.",
            "examples": [
              "internal_error"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Field-level problems. Populated for validation errors only.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary, safe to display. Never contains internals.",
            "examples": [
              "An unexpected error occurred."
            ],
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Correlation id for this request; quote it when reporting a problem.",
            "examples": [
              "9f1c2b7a4e5d4f0aa1b2c3d4e5f60718"
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "A single field-level problem, used for request-validation failures.",
        "properties": {
          "field": {
            "description": "Dotted path to the offending input, e.g. 'body.origin.lat'.",
            "examples": [
              "query.zoom"
            ],
            "title": "Field",
            "type": "string"
          },
          "message": {
            "description": "Human-readable explanation of why this input was rejected.",
            "examples": [
              "Input should be less than or equal to 22"
            ],
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "GeocodeMatch": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What the provider called this — a road, a building, a suburb.",
            "title": "Category"
          },
          "label": {
            "description": "Human-readable description, as the provider wrote it.",
            "title": "Label",
            "type": "string"
          },
          "latitude": {
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "title": "Longitude",
            "type": "number"
          }
        },
        "required": [
          "label",
          "longitude",
          "latitude"
        ],
        "title": "GeocodeMatch",
        "type": "object"
      },
      "GeocodeRequest": {
        "description": "Look up a place name within a pilot region.",
        "examples": [
          {
            "query": "Waterloo Public Square",
            "region": "waterloo"
          }
        ],
        "properties": {
          "limit": {
            "default": 5,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "query": {
            "description": "Place name or address to look for.",
            "maxLength": 160,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "region": {
            "default": "waterloo",
            "description": "Pilot region slug; results are restricted to its extent.",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z0-9-]+$",
            "title": "Region",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "GeocodeRequest",
        "type": "object"
      },
      "GeocodeResponse": {
        "properties": {
          "attribution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Credit the provider's licence requires.",
            "title": "Attribution"
          },
          "enabled": {
            "description": "False when no geocoder is configured. An empty `matches` with `enabled: false` means nothing was searched, not that nothing was found.",
            "title": "Enabled",
            "type": "boolean"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/GeocodeMatch"
            },
            "title": "Matches",
            "type": "array"
          },
          "provider": {
            "description": "Which geocoder answered, or 'disabled'.",
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "enabled",
          "matches"
        ],
        "title": "GeocodeResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend for **PathAble AI**, an accessibility-aware pedestrian routing project.\n\nRouting compares the shortest walking route against a route that respects a chosen\nmobility profile, and explains the difference using attributes recorded in\nOpenStreetMap.\n\n**What this is not.** Every routing decision is a deterministic rule over recorded\nmap attributes. There is no machine learning, no model prediction and no inferred\naccessibility score anywhere in this service — `ml_predictions_used` is present on\nevery route response and is always `false`. Missing accessibility data is reported\nas `unknown`, never as evidence that a path is clear, and no route is a guarantee\nthat a journey is passable.\n\nMap data © OpenStreetMap contributors, ODbL 1.0.",
    "title": "PathAble AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/geocode/search": {
      "post": {
        "description": "Searches for a place within a pilot region's extent. Submit-only: there is no as-you-type endpoint, because per-keystroke queries against a donated geocoding service are forbidden by its usage policy. Nothing about the request is stored.",
        "operationId": "searchPlaces",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeocodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeocodeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "The region does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The request did not match the expected schema."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unexpected server error. Quote the request id when reporting."
          },
          "503": {
            "description": "The geocoding provider failed."
          }
        },
        "summary": "Find coordinates for a place name",
        "tags": [
          "geocoding"
        ]
      }
    },
...
  },
...
}